Final Fixes to DAE
This is it. The penultimate set of fixes to the DAE class. In fact, it
is now fully implemented; the only problem is it leaks like crazy, which
will probably be the last thing I change in it.

All of the models from ORAS, packaged with textures and materials, are
coming your way. Next stop: Skeletal animations...