The Batch from Batch and More
The program can now be called from the command line, but it can
currently only do one file at a time in this manner.

Also, check for a "Body Mask" material present in certain Megas. Discard
them, because matching them to a texture would necessitate rewriting
most of the code I just pulled an all-nighter tweaking..