Exporting: Now with Fire
The DAE class now exports fire meshes correctly. Pokémon such as Mega
Charizard X, who have meshes around the mouth and on the tail that
simulate fire, now have properly-textured, transparent fire.

However, like the rest of the model, the fire meshes are not animated
and do not glow. You'll have to do that yourself when rendering the
model.